Major update to Reference, handles cues separate.
Instead of trying to recreate ranges of cues in the pars array, which is what ultimately drives the output when rendering back to a string, the Reference class has a *separate* cues property that is an array of the exact cues as calculated based on the pars.

Thus unlike this project's sibling Scripture library, which cleanly writes all verses to a verses array and thus reformats the scripture output string to a cleanly formatted string, THIS library will keep some of the original string formatting. In the par list, any single pars stay single pars, and a range stays a range, altho they are lightly beautified (capitalization, spacing). Other than that, they remain the same, but cues are then calculated as the more precise representation of the pars array, given that available cues are passed in with the availableCues property.

This also handles logic to look beyond a given paragraph's cues, knowing that scriptures may be following the paragraph but should be included. So common cues are searched for after the last matching paragraph cue, and cues will kept being added on until one of these "paragraph boundary" cues are found.