Add support for date-based publications. See #2.
Namely, Watchtower, Yearbook, Kingdom Ministry, and Meeting Workbooks.

Did it without needing to add new classes!

Added hasDate and date properties to the Publication class, then made it intelligent to format a Publication that has dates.

Correspondingly, the ReferenceUtil class had to be date-aware in its getPublication() method. Instead of just pulling a pre-instantiated publication in the list of publications, date-based publications are instantiated on-the-fly and have the date of the matched item passed into them. Special regex for retrieving the date in the format for the particular publication had to be provided, since different pubs reference their dates differently.

What's more, there's special consideration for Watchtower since it's formatting changed in 2016.