Improved support for mwb.
It uses "#1" convention instead of "P 1" convention, so adding support for that.

Had to update the PARBOUNDARY regex too, so that an individual cue could be grabbed.

Note there is a bug (which a unit test exposes) where chapter zero cannot be retrieved. This is typically used for first-week presentation videos.